Refactor playSimilar and add helper
