+++ ui/src/common/SongContextMenu.jsx
@@ -12,7 +12,7 @@ import { makeStyles } from '@material-ui/core/styles'
 import MoreVertIcon from '@material-ui/icons/MoreVert'
 import { MdQuestionMark } from 'react-icons/md'
 import clsx from 'clsx'
-import subsonic from '../subsonic'
+import { playSimilar } from '../utils'
 import {
   playNext,
   addTracks,
@@ -91,27 +91,8 @@ export const SongContextMenu = ({
     playSimilar: {
       enabled: true,
       label: translate('resources.song.actions.playSimilar'),
-      action: (record) => {
-        const id = record.mediaFileId || record.id
-        subsonic
-          .getSimilarSongs2(id, 100)
-          .then((res) => res.json['subsonic-response'])
-          .then((data) => {
-            if (data.status === 'ok') {
-              const songs = data.similarSongs2.song || []
-              const songData = {}
-              const ids = []
-              songs.forEach((s) => {
-                songData[s.id] = s
-                ids.push(s.id)
-              })
-              dispatch(playTracks(songData, ids))
-            }
-          })
-          .catch(() => {
-            notify('ra.page.error', 'warning')
-          })
-      },
+      action: (record) =>
+        playSimilar(dispatch, notify, record.mediaFileId || record.id),
     },
     addToPlaylist: {
       enabled: true,
