+++ ui/src/utils/index.js
@@ -3,3 +3,4 @@ export * from './intersperse'
 export * from './notifications'
 export * from './openInNewTab'
 export * from './urls'
+export * from './playSimilar'
